refactor: damage logic is joib specific resource
