+++ simaple/container/simulation.py
@@ -8,7 +8,7 @@ from simaple.core import ExtendedStat, JobCategory, JobType
 from simaple.data import get_best_ability
 from simaple.data.baseline import get_baseline_gearset
 from simaple.data.builtin_strategy import get_builtin_strategy
-from simaple.data import get_damage_logic
+from simaple.data.jobs.builtin import get_damage_logic
 from simaple.data.doping import get_normal_doping
 from simaple.data.passive import get_passive
 from simaple.data.jobs import get_skill_profile
