+++ simaple/data/__init__.py
@@ -1,8 +1,6 @@
-from simaple.data._builtin import get_damage_logic
 from simaple.data.ability import get_best_ability
 
 
 __all__ = [
-    "get_damage_logic",
     "get_best_ability",
 ]
