+++ simaple/data/jobs/resources/damage_logic.yaml
